CU-864e9nnrp - fix CLI tests
